Change `main` value in package.json from index.js to gatsby-node.js
This line caused 'Unable to find plugin' errors, which broke builds using the plugin
and prevented it working.

Changing the main value fixes the issue, but an alternative fix would be to remove the
'main' property entirely, and create an empty index.js file instead (as in many of the
official Gatsby plugins).